parser/requirements.txt
aiohttp
Pillow
numpy
fastapi
uvicorn
